Split out a module to be used by both the `aggregate` API method, and by the aggregate custom scan.
